Updates to the GUI
I have updated the GUI to make it look more suited to the application.
I have also added some functionality to the buttons.
In the next commit I will sort the service implementation as it is throwing errors now so I have commented it out.
I will also need to add code to the delete buttons and text field.